Add GET request using parameters
