+++ src/main/java/com/p1g14/pomodoro_timer_api/Greeting.java
@@ -1,0 +1,3 @@
+package com.p1g14.pomodoro_timer_api;
+
+public record Greeting(long id, String content) { }
