+++ src/main/java/com/p1g14/pomodoro_timer_api/GreetingController.java
@@ -1,0 +1,28 @@
+package com.p1g14.pomodoro_timer_api;
+
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+@RestController
+public class GreetingController {
+    private static final String template = "Hello, %s!";
+    private final AtomicLong counter = new AtomicLong();
+
+    @GetMapping("/")
+    public String hello() {
+        return "Hello world!";
+    }
+
+    @GetMapping("/user")
+    public String helloUser() {
+        return "Hello user!";
+    }
+    @GetMapping("/greeting")
+    public Greeting greeting(@RequestParam(value = "name", defaultValue = "World") String name) {
+        return new Greeting(counter.incrementAndGet(), String.format(template, name));
+    }
+}
