+++ src/main/java/com/p1g14/pomodoro_timer_api/PomodoroTimerApiApplication.java
@@ -6,19 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
-@RestController
 public class PomodoroTimerApiApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(PomodoroTimerApiApplication.class, args);
-	}
-
-	@GetMapping("/")
-	public String hello() {
-		return String.format("Hello world!");
-	}
-
-	@GetMapping("/user")
-	public String helloUser() {
-		return String.format("Hello user!");
 	}
 }
